put in the class structure (c.f. discussion on Oct 27)
