+++ source/FullOrderModel.py
@@ -1,0 +1,17 @@
+class FullOrderModel:
+    """! Full-order-model (FOM)
+    This is the general parent class for full-order models.
+    For any particular model the user should create a subclass and specify the functions below.
+    """
+
+    def __init__(self):
+        """!
+        the main part of the initialization is entirely up to the user. They should overwrite the __init__ call with
+        what they need. In case this call will at some point include some reference setup, the user can still call it
+        with super().__init__()
+        """
+        # TODO: set the important variables
+        pass
+
+    # TODO: create those functions that we know we'll need, add explanations for their correct setup
+
