+++ source/leftover/Brain.py
