+++ source/leftover/NoiseModel.py
